refactor(refund-fees): update refund fee data handling and remove unused components
