+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-fees/[id]/form.tsx
@@ -1,16 +1,29 @@
 "use client";
 import { Button } from "@/components/ui/button";
 import { toast } from "@/components/ui/sonner";
-import type { UniRefund_ContractService_Refunds_RefundFeeHeaders_RefundFeeHeaderDto } from "@ayasofyazilim/saas/ContractService";
-import { $UniRefund_ContractService_Refunds_RefundFeeHeaders_RefundFeeHeaderUpdateDto } from "@ayasofyazilim/saas/ContractService";
+import type {
+  UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto,
+  UniRefund_ContractService_Refunds_RefundFeeHeaders_RefundFeeHeaderDto,
+} from "@ayasofyazilim/saas/ContractService";
+import {
+  $UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto,
+  $UniRefund_ContractService_Refunds_RefundFeeHeaders_RefundFeeHeaderUpdateDto,
+} from "@ayasofyazilim/saas/ContractService";
 import ConfirmDialog from "@repo/ayasofyazilim-ui/molecules/confirm-dialog";
+import { tanstackTableEditableColumnsByRowData } from "@repo/ayasofyazilim-ui/molecules/tanstack-table/utils";
 import { SchemaForm } from "@repo/ayasofyazilim-ui/organisms/schema-form";
+import { TableField } from "@repo/ayasofyazilim-ui/organisms/schema-form/fields";
 import { createUiSchemaWithResource } from "@repo/ayasofyazilim-ui/organisms/schema-form/utils";
 import { ActionList } from "@repo/ui/action-button";
+import { PlusCircle, Trash2 } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { handlePutResponse } from "src/actions/core/api-utils-client";
-import { putRefundFeeHeadersApi } from "src/actions/unirefund/ContractService/put-actions";
+import { putRefundFeeHeadersByIdApi } from "src/actions/unirefund/ContractService/put-actions";
 import type { ContractServiceResource } from "src/language-data/unirefund/ContractService";
+
+type TypeWithId<Type, IdType = string> = Type & {
+  id: IdType;
+};
 
 function Form({
   response,
@@ -20,21 +33,93 @@ function Form({
   languageData: ContractServiceResource;
 }) {
   const router = useRouter();
+  const RebateFeeColumns = tanstackTableEditableColumnsByRowData<
+    TypeWithId<UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto>
+  >({
+    rows: {
+      ...$UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto.properties,
+      feeType: {
+        ...$UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto
+          .properties.feeType,
+        enum: $UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto.properties.feeType.enum.map(
+          (item) => ({
+            value: item,
+            label: item,
+          }),
+        ),
+      },
+      refundMethod: {
+        ...$UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto
+          .properties.refundMethod,
+        enum: $UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto.properties.refundMethod.enum.map(
+          (item) => ({
+            value: item,
+            label: item,
+          }),
+        ),
+      },
+    },
+  });
   const uiSchema = createUiSchemaWithResource({
     resources: languageData,
     schema:
       $UniRefund_ContractService_Refunds_RefundFeeHeaders_RefundFeeHeaderUpdateDto,
+    extend: {
+      refundFeeDetails: {
+        "ui:field": "RefundFee",
+      },
+    },
   });
   return (
     <SchemaForm
-      className="h-auto"
+      className="w-full"
+      fields={{
+        RefundFee: TableField<
+          TypeWithId<UniRefund_ContractService_Refunds_RefundFeeDetails_RefundFeeDetailDto>
+        >({
+          editable: true,
+          columns: RebateFeeColumns,
+          data: response.refundFeeDetails || [],
+          fillerColumn: "id",
+          tableActions: [
+            {
+              type: "create-row",
+              actionLocation: "table",
+              cta: languageData["Rebate.Create"],
+              icon: PlusCircle,
+            },
+          ],
+          rowActions: [
+            {
+              actionLocation: "row",
+              cta: languageData.Delete,
+              icon: Trash2,
+              type: "delete-row",
+            },
+          ],
+          columnVisibility: {
+            type: "hide",
+            columns: [
+              "id",
+              "creationTime",
+              "creatorId",
+              "lastModificationTime",
+              "lastModifierId",
+              "isDeleted",
+              "deleterId",
+              "deletionTime",
+              "refundFeeHeaderId",
+            ],
+          },
+        }),
+      }}
       formData={response}
       onSubmit={(data) => {
         const formData = {
           id: response.id,
           requestBody: data.formData,
         };
-        void putRefundFeeHeadersApi(formData).then((res) => {
+        void putRefundFeeHeadersByIdApi(formData).then((res) => {
           handlePutResponse(res, router);
         });
       }}
@@ -45,7 +130,7 @@ function Form({
       useDefaultSubmit={false}
       withScrollArea={false}
     >
-      <ActionList className="mb-4 border-0 p-0">
+      <ActionList className="sticky bottom-0 z-10 mb-4 border-0 bg-white p-0 py-4">
         <ConfirmDialog
           confirmProps={{
             onConfirm: () => {
