+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/refund-fees/[id]/page.tsx
@@ -4,7 +4,6 @@ import { getRefundFeeHeadersByIdApi } from "src/actions/unirefund/ContractServic
 import { isUnauthorized } from "src/utils/page-policy/page-policy";
 import { getResourceData } from "src/language-data/unirefund/ContractService";
 import Form from "./form";
-import RefundFeeDetailsForm from "./table";
 
 export default async function Page({
   params,
@@ -30,10 +29,6 @@ export default async function Page({
   return (
     <>
       <Form languageData={languageData} response={response.data} />
-      <RefundFeeDetailsForm
-        languageData={languageData}
-        response={response.data}
-      />
       <div className="hidden" id="page-title">
         {response.data.name}
       </div>
